requirements.txt
openai>=1.30.0
sounddevice>=0.4.6
numpy>=1.26.0
pynput>=1.7.6
python-dotenv>=1.0.0
pystray>=0.19.5
Pillow>=10.0.0
